backend/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
numpy
pandas
google-genai
dotenv